apps/user-service/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
redis>=5.0.0,<6.0.0
prometheus-fastapi-instrumentator>=7.0.0,<8.0.0
pydantic>=2.0.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
